Add authentication, authorization and user CRUD tests

+++ src/main/java/hexlet/code/controller/api/UsersController.java
@@ -4,10 +4,12 @@ import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserUpdateDTO;
 import hexlet.code.service.UserService;
+import hexlet.code.util.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -16,6 +18,8 @@ import java.util.List;
 public class UsersController {
     @Autowired
     private UserService userService;
+    @Autowired
+    private UserUtils userUtils;
 
     @GetMapping("")
     @ResponseStatus(HttpStatus.OK)
@@ -38,13 +42,20 @@ public class UsersController {
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public UserDTO update(@Valid @RequestBody UserUpdateDTO userData, @PathVariable Long id) {
+        var currentUser = userUtils.getCurrentUser();
+        if (!currentUser.getId().equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         return userService.update(userData, id);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void destroy(@PathVariable Long id) {
+        var currentUser = userUtils.getCurrentUser();
+        if (!currentUser.getId().equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         userService.delete(id);
     }
-
 }
